[philosophers] time check comparing with reference. TODO: solve problem about printing delay.

+++ cursus/philosophers/philo_one/get_cur_time.c
@@ -8,6 +8,6 @@ uint64_t
 
 	if (gettimeofday(&tmp, NULL) == -1)
 		return (-1);
-	ret = tmp.tv_sec * SEC2USEC + tmp.tv_usec * USEC2MSEC;
+	ret = tmp.tv_sec * (uint64_t)SEC2USEC + tmp.tv_usec * USEC2MSEC;
 	return (ret);
 }
